% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto_data_utils.R
\name{crypto_get_binance_coin_m_perps_funding}
\alias{crypto_get_binance_coin_m_perps_funding}
\title{Load Binance Coin-M Perp funding data}
\usage{
crypto_get_binance_coin_m_perps_funding(path = "binance", force_update = TRUE)
}
\arguments{
\item{path}{The path to save the dataset locally.}

\item{force_update}{Force download and overwrite exsiting files}
}
\value{
The binance coin-M perp funding dataset as a tibble.
}
\description{
Load Binance Coin-M Perp funding data
}
\examples{
\dontrun{
df <- crypto_get_binance_coin_m_perps_funding()
}
}
