% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto_data_utils.R
\name{crypto_get_binance_perps_1h}
\alias{crypto_get_binance_perps_1h}
\title{Load Binance Perp 1h OHLCV data}
\usage{
crypto_get_binance_perps_1h(path = "binance", force_update = TRUE)
}
\arguments{
\item{path}{The path to save the dataset locally.}

\item{force_update}{Force download and overwrite exsiting files}
}
\value{
The binance 1h perp OHLCV dataset as a tibble.
}
\description{
Load Binance Perp 1h OHLCV data
}
\examples{
\dontrun{
df <- crypto_get_binance_perps_1h()
}
}
