% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto_data_utils.R
\name{crypto_get_clean_spot}
\alias{crypto_get_clean_spot}
\title{Load cleaned spot hourly data from FTX
This dataset is cleaned from tokenzied stocks, leveraged tokens and FX pairs (only includes USDC denominated pairs)}
\usage{
crypto_get_clean_spot(path = "ftx", force_update = TRUE)
}
\arguments{
\item{path}{The path to save the dataset locally.}

\item{force_update}{Force download and overwrite exsiting files}
}
\value{
The spot dataset as a tibble.
}
\description{
Load cleaned spot hourly data from FTX
This dataset is cleaned from tokenzied stocks, leveraged tokens and FX pairs (only includes USDC denominated pairs)
}
\examples{
\dontrun{
df <- crypto_get_clean_spot()
}
}
