% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto_data_utils.R
\name{crypto_get_coincodex}
\alias{crypto_get_coincodex}
\title{Load coincodex daily crypto data}
\usage{
crypto_get_coincodex(path = "coincodex", force_update = TRUE)
}
\arguments{
\item{path}{The path to save the coincodex dataset locally.}

\item{force_update}{Force download and overwrite exsiting files}
}
\value{
The coincodex dataset as a tibble.
}
\description{
Load coincodex daily crypto data
}
\examples{
\dontrun{
df <- crypto_get_coincodex()
}
}
