% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto_data_utils.R
\name{crypto_get_coinmetrics}
\alias{crypto_get_coinmetrics}
\title{Load coinmetrics daily crypto data}
\usage{
crypto_get_coinmetrics(path = "coinmetrics", force_update = TRUE)
}
\arguments{
\item{path}{The path to save the Coinmetrics dataset locally.}

\item{force_update}{Force download and overwrite exsiting files}
}
\value{
The Coinmetrics dataset as a tibble.
}
\description{
Load coinmetrics daily crypto data
}
\examples{
\dontrun{
df <- crypto_get_coinmetrics()
}
}
