% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto_data_utils.R
\name{crypto_get_futures}
\alias{crypto_get_futures}
\title{Load futures hourly data from FTX}
\usage{
crypto_get_futures(path = "ftx", force_update = TRUE)
}
\arguments{
\item{path}{The path to save the dataset locally.}

\item{force_update}{Force download and overwrite exsiting files}
}
\value{
The futures dataset as a tibble.
}
\description{
Load futures hourly data from FTX
}
\examples{
\dontrun{
df <- crypto_get_futures()
}
}
