% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto_data_utils.R
\name{crypto_get_minute_perpetuals}
\alias{crypto_get_minute_perpetuals}
\title{Load One minnute perpetuals data from FTX}
\usage{
crypto_get_minute_perpetuals(path = "ftx", force_update = TRUE)
}
\arguments{
\item{path}{The path to save the dataset locally.}

\item{force_update}{Force download and overwrite exsiting files}
}
\value{
The one minute perpetuals dataset as a tibble.
}
\description{
Load One minnute perpetuals data from FTX
}
\examples{
\dontrun{
df <- crypto_get_minute_perpetuals()
}
}
