% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto_data_utils.R
\name{crypto_get_perp_rates}
\alias{crypto_get_perp_rates}
\title{Load Perpetual Funding Rates data from FTX}
\usage{
crypto_get_perp_rates(path = "ftx", force_update = TRUE)
}
\arguments{
\item{path}{The path to save the dataset locally.}

\item{force_update}{Force download and overwrite exsiting files}
}
\value{
The Perpetual Funding Rates dataset as a tibble.
}
\description{
Load Perpetual Funding Rates data from FTX
}
\examples{
\dontrun{
df <- crypto_get_perp_rates()
}
}
