% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto_data_utils.R
\name{crypto_get_rebalance_trades}
\alias{crypto_get_rebalance_trades}
\title{Load rebalance token data from FTX}
\usage{
crypto_get_rebalance_trades(path = "ftx", force_update = TRUE)
}
\arguments{
\item{path}{The path to save the dataset locally.}

\item{force_update}{Force download and overwrite exsiting files}
}
\value{
The rebalance trades dataset as a tibble.
}
\description{
Load rebalance token data from FTX
}
\examples{
\dontrun{
df <- crypto_get_rebalance_trades()
}
}
