% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto_data_utils.R
\name{crypto_get_spot}
\alias{crypto_get_spot}
\title{Load spot hourly data from FTX}
\usage{
crypto_get_spot(path = "ftx", force_update = TRUE)
}
\arguments{
\item{path}{The path to save the dataset locally.}

\item{force_update}{Force download and overwrite exsiting files}
}
\value{
The spot dataset as a tibble.
}
\description{
Load spot hourly data from FTX
}
\examples{
\dontrun{
df <- crypto_get_spot()
}
}
