% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto_data_utils.R
\name{crypto_get_trades_sample}
\alias{crypto_get_trades_sample}
\title{Load trades sample for BTC-PERP from FTX}
\usage{
crypto_get_trades_sample(path = "ftx", force_update = TRUE)
}
\arguments{
\item{path}{The path to save the dataset locally.}

\item{force_update}{Force download and overwrite exsiting files}
}
\value{
The trades snapshot dataset as a tibble.
}
\description{
Load trades sample for BTC-PERP from FTX
}
\examples{
\dontrun{
df <- crypto_get_trades_sample()
}
}
