% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equity_factors_utils.R
\name{equity_get_liquid_universe}
\alias{equity_get_liquid_universe}
\title{Load equity factors liquid universe data}
\usage{
equity_get_liquid_universe(path = "equityfactors", force_update = TRUE)
}
\arguments{
\item{path}{The path to save the dataset locally.}

\item{force_update}{Force download and overwrite existing files}
}
\value{
The dataset as a data frame.
}
\description{
Load equity factors liquid universe data
}
\examples{
\dontrun{
df <- equity_get_liquid_universe()
}
}
