% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equity_factors_utils.R
\name{equity_get_statarb_acquisitions}
\alias{equity_get_statarb_acquisitions}
\title{Load stat arb acquisitions data}
\usage{
equity_get_statarb_acquisitions(path = "equityfactors", force_update = TRUE)
}
\arguments{
\item{path}{The path to save the dataset locally.}

\item{force_update}{Force download and overwrite existing files}
}
\value{
The stat arb acquisitions dataset as a data frame.
}
\description{
Load stat arb acquisitions data
}
\examples{
\dontrun{
df <- equity_get_statarb_acquisitions()
}
}
