% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equity_factors_utils.R
\name{equity_get_statarb_prices}
\alias{equity_get_statarb_prices}
\title{Load stat arb prices data}
\usage{
equity_get_statarb_prices(path = "equityfactors", force_update = TRUE)
}
\arguments{
\item{path}{The path to save the dataset locally.}

\item{force_update}{Force download and overwrite existing files}
}
\value{
The stat arb prices dataset as a data frame.
}
\description{
Load stat arb prices data
}
\examples{
\dontrun{
df <- equity_get_statarb_prices()
}
}
