% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equity_factors_utils.R
\name{equity_get_statarb_spreads}
\alias{equity_get_statarb_spreads}
\title{Load stat arb spreads data}
\usage{
equity_get_statarb_spreads(path = "equityfactors", force_update = TRUE)
}
\arguments{
\item{path}{The path to save the dataset locally.}

\item{force_update}{Force download and overwrite existing files}
}
\value{
The stat arb spreads dataset as a data frame.
}
\description{
Load stat arb spreads data
}
\examples{
\dontrun{
df <- equity_get_statarb_spreads()
}
}
