% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_data_utils.R
\name{fx_get_asset_list}
\alias{fx_get_asset_list}
\title{Load a Zorro Assets List}
\usage{
fx_get_asset_list(asset_list, path = "Zorro-Assets-Lists", force_update = TRUE)
}
\arguments{
\item{asset_list}{The name of the Zorro Asset List to load, without the
file extension.}

\item{path}{The path to save the Asset List locally.}

\item{force_update}{Force download and overwrite exsiting files}
}
\value{
The Asset List as a tibble.
}
\description{
If the Assets List file does not exist in \code{path}, download it from
the Research Pod's GCS bucket.
}
\examples{
\dontrun{
fx_get_asset_list("AssetsDWX-FX-USD")
}
}
