% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_data_utils.R
\name{fx_get_daily_OHLC_ticker}
\alias{fx_get_daily_OHLC_ticker}
\title{Load historical daily price data of a ticker}
\usage{
fx_get_daily_OHLC_ticker(ticker, path = "Daily", force_update = TRUE)
}
\arguments{
\item{ticker}{The ticker to load.}

\item{path}{The path to save the daily resolution price data.}

\item{force_update}{Force download and overwrite exsiting files}
}
\value{
A data.frame of market data.
}
\description{
Returns a data.frame of market data with columns
Ticker, Date, Open, High, Low, Close, Volume
}
\examples{
\dontrun{
fx_get_daily_OHLC_ticker("EURUSD")
}
}
