% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_data_utils.R
\name{fx_get_unique_currencies}
\alias{fx_get_unique_currencies}
\title{Unique currencies from fx pairs}
\usage{
fx_get_unique_currencies(prices_df)
}
\arguments{
\item{prices_df}{A data.frame of fx price data}
}
\value{
A chr vector of unique currencies
}
\description{
Get unique currencies from a data frame of price data.
}
\examples{
\dontrun{
fx_get_unique_currencies(prices_df)
}
}
