% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fx_data_utils.R
\name{fx_total_return_index}
\alias{fx_total_return_index}
\title{Calculate currency total return indexes}
\usage{
fx_total_return_index(prices_df, policy_rates_df)
}
\arguments{
\item{prices_df}{A data.frame of historical spot prices}

\item{policy_rates_df}{A data.frame of historical central bank policy rates}
}
\value{
A data.frame of index returns: spot, interest and total
}
\description{
Records without corresponding rate data are removed from returned
data.frame.
}
