% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{get_lab_app}
\alias{get_lab_app}
\title{Create oauth app for use in the lab}
\usage{
get_lab_app()
}
\value{
oauth app
}
\description{
Surprisingly, the “client secret” in an oauth_app() is not a secret. It’s not equivalent to a password, and if you are writing an API wrapper package, it should be included in the package. \href{https://developers.google.com/identity/protocols/OAuth2#installed}{(If you don’t believe me, here are google’s comments on the topic.)}
}
