% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwlab_gcs.R
\name{get_pod_meta}
\alias{get_pod_meta}
\title{Get Research Pod metadata}
\usage{
get_pod_meta(pod = NA)
}
\arguments{
\item{pod}{string, The name of the Research Pod. If NA, data for all pods is returned.}
}
\value{
List of metadata for the specified Research Pod, or list of lists of metadata for all Pods if \code{pod} is NA.
}
\description{
Returns bucket and dataset names for a specified pod,
or all pods if \code{pod} is unspecified.
}
\examples{
bucket = get_pod_meta("EquityFactors")[["bucket"]]
datasets = get_pod_meta("EquityFactors")[["datasets"]]
all_pods_meta = get_pod_meta()
}
