% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwlab_gcs.R
\name{list_pods}
\alias{list_pods}
\title{Get the names of The Lab's Research Pods}
\usage{
list_pods()
}
\value{
A character vector of the names of The Lab's Research Pods
}
\description{
Get the names of The Lab's Research Pods
}
\examples{
list_pods()
}
