% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwlab_gcs.R
\name{load_pod_data}
\alias{load_pod_data}
\title{Transfer all data relevant to a specific Research Pod from Google Cloud and save to disk}
\usage{
load_pod_data(pod, path = ".")
}
\arguments{
\item{pod}{string, The name of the research pod. Options: "Equity Factors" (others TBA)}

\item{path}{string, Path to the folder where data will be saved. Defaults to root}
}
\value{
None, transfers data from Google Cloud to \code{path}
}
\description{
Requires authorisation. Set up prior with \code{rwlab_gc_auth}.
}
\examples{
\dontrun{
rwlab_gc_auth("your-email-address")
load_pod_data("Equity Factors", "./data")
}
}
