% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_pod_utils.R
\name{macro_get_close_price_momo}
\alias{macro_get_close_price_momo}
\title{Load stock closing price momentum data - prices snapshotted 10-15 minutes
before the close, and the close.}
\usage{
macro_get_close_price_momo(path = "macropod", force_update = TRUE)
}
\arguments{
\item{path}{The path to save the dataset locally.}

\item{force_update}{Force download and overwrite exsiting files}
}
\value{
The US stock closing price momentum dataset as a tibble.
}
\description{
Load stock closing price momentum data - prices snapshotted 10-15 minutes
before the close, and the close.
}
\examples{
\dontrun{
df <- macro_get_close_price_momo()
}
}
