% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_pod_utils.R
\name{macro_get_expiring_futures}
\alias{macro_get_expiring_futures}
\title{Load expiring futures daily OHLC}
\usage{
macro_get_expiring_futures(path = "macropod", force_update = TRUE)
}
\arguments{
\item{path}{The path to save the dataset locally.}

\item{force_update}{Force download and overwrite exsiting files}
}
\value{
The futures dataset as a tibble.
}
\description{
Load expiring futures daily OHLC
}
\examples{
\dontrun{
df <- macro_get_expiring_futures()
}
}
