% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_pod_utils.R
\name{macro_get_expiring_vx_futures}
\alias{macro_get_expiring_vx_futures}
\title{Load expiring VX futures daily OHLC plus metadata}
\usage{
macro_get_expiring_vx_futures(path = "macropod", force_update = TRUE)
}
\arguments{
\item{path}{The path to save the dataset locally.}

\item{force_update}{Force download and overwrite exsiting files}
}
\value{
The VX futures dataset as a tibble.
}
\description{
Load expiring VX futures daily OHLC plus metadata
}
\examples{
\dontrun{
df <- macro_get_expiring_vx_futures()
}
}
