% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_pod_utils.R
\name{macro_get_govt_bond_returns}
\alias{macro_get_govt_bond_returns}
\title{Load government bond returns dataset}
\usage{
macro_get_govt_bond_returns(path = "macropod", force_update = TRUE)
}
\arguments{
\item{path}{The path to save the dataset locally.}

\item{force_update}{Force download and overwrite existing files}
}
\value{
The govt bonds returns dataset as a tibble.
}
\description{
Load government bond returns dataset
}
\examples{
\dontrun{
df <- macro_get_govt_bond_returns()
}
}
