% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_pod_utils.R
\name{macro_get_historical_asset_class}
\alias{macro_get_historical_asset_class}
\title{Load main asset classes daily  data}
\usage{
macro_get_historical_asset_class(path = "macropod", force_update = TRUE)
}
\arguments{
\item{path}{The path to save the dataset locally.}

\item{force_update}{Force download and overwrite exsiting files}
}
\value{
Historical daily data of some of the main asset classes as a tibble.
}
\description{
Load main asset classes daily  data
}
\examples{
\dontrun{
df <- macro_get_historical_asset_class()
}
}
