% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_pod_utils.R
\name{macro_get_historical_short_sale}
\alias{macro_get_historical_short_sale}
\title{Load Interactive Brokers historical short sale data}
\usage{
macro_get_historical_short_sale(
  year = 2023,
  path = "macropod",
  force_update = TRUE
)
}
\arguments{
\item{path}{The path to save the dataset locally.}

\item{force_update}{Force download and overwrite exsiting files}
}
\value{
Historical daily short sale data as a tibble.
}
\description{
Load Interactive Brokers historical short sale data
}
\examples{
\dontrun{
df <- macro_get_historical_short_sale()
}
}
