% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_pod_utils.R
\name{macro_get_nyse_holidays}
\alias{macro_get_nyse_holidays}
\title{Load NYSE holidays}
\usage{
macro_get_nyse_holidays(path = "macropod", force_update = TRUE)
}
\arguments{
\item{path}{The path to save the dataset locally.}

\item{force_update}{Force download and overwrite exsiting files}
}
\value{
NYSE holidays from 1990 through 2040.
}
\description{
Load NYSE holidays
}
\examples{
\dontrun{
df <- macro_get_nyse_holidays()
}
}
