% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_pod_utils.R
\name{macro_get_straddles_over_earnings}
\alias{macro_get_straddles_over_earnings}
\title{Load straddles over earnings data}
\usage{
macro_get_straddles_over_earnings(path = "macropod", force_update = TRUE)
}
\arguments{
\item{path}{The path to save the dataset locally.}

\item{force_update}{Force download and overwrite exsiting files}
}
\value{
The straddles over earnings dataset as a tibble.
}
\description{
Load straddles over earnings data
}
\examples{
\dontrun{
df <- macro_get_straddles_over_earnings()
}
}
