% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_pod_utils.R
\name{macro_get_vix_vix3m}
\alias{macro_get_vix_vix3m}
\title{Load VIX and VIX3M index daily  data}
\usage{
macro_get_vix_vix3m(path = "macropod", force_update = TRUE)
}
\arguments{
\item{path}{The path to save the vix/vix3m dataset locally.}

\item{force_update}{Force download and overwrite exsiting files}
}
\value{
Historical daily index data of VIX and VIX3M as a tibble.
}
\description{
Load VIX and VIX3M index daily  data
}
\examples{
\dontrun{
df <- macro_get_vix_vix3m()
}
}
