% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwlab_gcs.R
\name{quicksetup}
\alias{quicksetup}
\title{Transfers prices data from Research Pod data library and adds it as a \code{data.frame} to the Global Environment}
\usage{
quicksetup(pod, path = ".")
}
\arguments{
\item{pod}{string, The name of the Research Pod}

\item{path}{string, The path to the local directory to save the transferred data}
}
\value{
nothing, but saves Research Pod price data locally and populates Global Env with a \code{data.frame}
}
\description{
Transfers prices data from Research Pod data library and adds it as a \code{data.frame} to the Global Environment
}
\examples{
\dontrun{
quickset("EquityFactors")
}
}
