% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwlab_gcs.R
\name{rwlab_list_bucket_objects}
\alias{rwlab_list_bucket_objects}
\title{List objects in a RW Lab Google Cloud bucket}
\usage{
rwlab_list_bucket_objects(bucket = "rw_equity_research_sprint")
}
\arguments{
\item{bucket}{string, The name of the bucket}
}
\value{
A list of objects in \code{bucket}
}
\description{
Requires authorisation. Set up prior with \code{rwlab_gc_auth}.
}
\examples{
\dontrun{
rwlab_gc_auth("your-email-address")
rwlab_list_bucket_objects("rw_equity_research_sprint")
}
}
