% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwlab_gcs.R
\name{setup_for_pod}
\alias{setup_for_pod}
\title{Setup for a Research Pod}
\usage{
setup_for_pod(pod, path = ".", oauth_email = NA)
}
\arguments{
\item{pod}{string, The Research Pod to set up}

\item{path}{string, The local path to use to save Pod data}

\item{oauth_email}{string, optional, The email to authorise to cloud storage, default NA. NA requires interactively selecting an email}
}
\value{
bool specifying success (TRUE) or failure (FALSE) of Pod object transfers
}
\description{
Authorise to cloud storage and transfer data for a Research Pod
}
\examples{
\dontrun{
setup_for_pod("EquityFactors")
}
}
