% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwlab_gcs.R
\name{transfer_pod_data}
\alias{transfer_pod_data}
\title{Transfer all Research Pod data from Google Cloud to disk}
\usage{
transfer_pod_data(pod, path = ".")
}
\arguments{
\item{pod}{string, The name of the research pod. Options: "Equity Factors" (others TBA)}

\item{path}{string, Path to the folder where data will be saved. Defaults to root}
}
\value{
bool specifying success (TRUE) or failure (FALSE) of object transfers. Returns FALSE on first failure.
}
\description{
Requires authorisation. Set up prior with \code{rwlab_gc_auth}.
}
\examples{
\dontrun{
rwlab_gc_auth("your-email-address")
transfer_pod_data("Equity Factors", "./data")
}
}
