% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwlab_gcs.R
\name{load_lab_object}
\alias{load_lab_object}
\title{Transfer object from data library and return it as a \code{data.frame}}
\usage{
load_lab_object(pod, object, path = ".")
}
\arguments{
\item{pod}{string, The name of the Research Pod}

\item{object}{string, The name of the data object}

\item{path}{string, The local path to the directory to save the data object. defaults to "."}
}
\value{
\code{object} as a \code{data.frame}
}
\description{
Transfer object from data library and return it as a \code{data.frame}
}
\examples{
\dontrun{
load_lab_object("EquityFactors", "R1000_fundamentals_1d.feather")
}
}
