% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwlab_gcs.R
\name{load_lab_object}
\alias{load_lab_object}
\title{Transfer a single specified object from Google Cloud}
\usage{
load_lab_object(path, object, bucket = NA)
}
\arguments{
\item{path}{string, Local path for saving object}

\item{object}{string, Name of the object to transfer}

\item{bucket}{string, Name of object's bucket, NA uses gcs_global_bucket, default NA}
}
\value{
bool specifying success (TRUE) or failure (FALSE) of object transfer
}
\description{
This is useful for reloading a specific object rather than transferring all Research Pod objects.
}
\examples{
\dontrun{
load_data_object(path = ".", object = "clean_R1000.csv", bucket = "rw_equity_research_sprint")
}
}
