% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwlab_lab_auth.R
\name{rwlab_data_auth}
\alias{rwlab_data_auth}
\title{Interactively authorise a Google Identity and load Oauth2 token}
\usage{
rwlab_data_auth(oauth_email = NA, oauth_cache = FALSE)
}
\arguments{
\item{oauth_email}{string, The email address you use to access The Lab. If NA (default), select interactively.}

\item{oauth_cache}{bool, Cache the OAuth token for reuse until expiry}
}
\value{
an OAuth token object (invisibly)
}
\description{
Authorise a Google Identity to access The Lab's cloud infrastructure. The first call
in a session asks the user to interactively specify a Google Identity and provide the
resulting code back to the call site.
}
\details{
Internally sets the following options:
\itemize{
\item \code{gargle::gargle_oauth_email}
\item \code{gargle::gargle_oath_cache}
\item \code{gargle::gargle_oob_default}
\item \code{rlang::rlang_interactive} (if called from an interactive session)
}

See \code{\link[gargle:gargle_options]{gargle_options}} for details.

TODO: How to set up a Google Identity if you don't have one and/or don't want to use a gmail account.
}
\examples{
\dontrun{rwlab_data_auth()}
}
