% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwlab_gcs.R
\name{transfer_lab_object}
\alias{transfer_lab_object}
\title{Transfer a single specified object from data library}
\usage{
transfer_lab_object(pod, object, path = ".")
}
\arguments{
\item{pod}{string, Name of the Research Pod}

\item{object}{string, Name of the object to transfer}

\item{path}{string, Local path for saving object, defaults to "."}
}
\value{
\code{bool} specifying success (TRUE) or failure (FALSE) of object transfer
}
\description{
This is useful for reloading a specific object rather than transferring all Research Pod objects.
}
\examples{
\dontrun{
load_lab_object(path = ".", object = "clean_R1000.csv", bucket = "rw_equity_research_sprint")
}
}
